Revert "[ENH] 1965 Split up embedding functions (#2034)"
This reverts commit a5ea9a5c6aa42a7f274b9445c5bfb9553c14aee8.
