+++ chromadb/test/ef/test_default_ef.py
@@ -7,10 +7,7 @@ import onnxruntime
 import pytest
 from hypothesis import given, settings
 
-from chromadb.utils.embedding_functions.onnx_mini_lm_l6_v2 import (
-    ONNXMiniLM_L6_V2,
-    _verify_sha256,
-)
+from chromadb.utils.embedding_functions import ONNXMiniLM_L6_V2, _verify_sha256
 
 
 def unique_by(x: Hashable) -> Hashable:
