--- chromadb/utils/embedding_functions/amazon_bedrock_embedding_function.py
@@ -1,55 +1,0 @@
-import json
-import logging
-from typing import Any
-
-from chromadb.api.types import Documents, EmbeddingFunction, Embeddings
-
-logger = logging.getLogger(__name__)
-
-
-class AmazonBedrockEmbeddingFunction(EmbeddingFunction[Documents]):
-    def __init__(
-        self,
-        session: Any,
-        model_name: str = "amazon.titan-embed-text-v1",
-        **kwargs: Any,
-    ):
-        """Initialize AmazonBedrockEmbeddingFunction.
-
-        Args:
-            session (boto3.Session): The boto3 session to use. You need to have boto3
-            installed, `pip install boto3`.
-            model_name (str, optional): Identifier of the model, defaults to "amazon.titan-embed-text-v1"
-            **kwargs: Additional arguments to pass to the boto3 client.
-
-        Example:
-            >>> import boto3
-            >>> session = boto3.Session(profile_name="profile", region_name="us-east-1")
-            >>> bedrock = AmazonBedrockEmbeddingFunction(session=session)
-            >>> texts = ["Hello, world!", "How are you?"]
-            >>> embeddings = bedrock(texts)
-        """
-
-        self._model_name = model_name
-
-        self._client = session.client(
-            service_name="bedrock-runtime",
-            **kwargs,
-        )
-
-    def __call__(self, input: Documents) -> Embeddings:
-        accept = "application/json"
-        content_type = "application/json"
-        embeddings = []
-        for text in input:
-            input_body = {"inputText": text}
-            body = json.dumps(input_body)
-            response = self._client.invoke_model(
-                body=body,
-                modelId=self._model_name,
-                accept=accept,
-                contentType=content_type,
-            )
-            embedding = json.load(response.get("body")).get("embedding")
-            embeddings.append(embedding)
-        return embeddings
