--- chromadb/utils/embedding_functions/cohere_embedding_function.py
@@ -1,27 +1,0 @@
-import logging
-
-from chromadb.api.types import Documents, EmbeddingFunction, Embeddings
-
-logger = logging.getLogger(__name__)
-
-
-class CohereEmbeddingFunction(EmbeddingFunction[Documents]):
-    def __init__(self, api_key: str, model_name: str = "large"):
-        try:
-            import cohere
-        except ImportError:
-            raise ValueError(
-                "The cohere python package is not installed. Please install it with `pip install cohere`"
-            )
-
-        self._client = cohere.Client(api_key)
-        self._model_name = model_name
-
-    def __call__(self, input: Documents) -> Embeddings:
-        # Call Cohere Embedding API for each document.
-        return [
-            embeddings
-            for embeddings in self._client.embed(
-                texts=input, model=self._model_name, input_type="search_document"
-            )
-        ]
